[Gold V] Title: 회의실 배정, Time: 516 ms, Memory: 46388 KB -BaekjoonHub

+++ 백준/Gold/1931. 회의실 배정/회의실 배정.java
@@ -8,19 +8,19 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
-        long[][] arr = new long[N][2];
+        int[][] arr = new int[N][2];
         for (int i = 0; i < N; i++) {
             StringTokenizer st = new StringTokenizer(br.readLine());
-            arr[i][0] = Long.parseLong(st.nextToken());
-            arr[i][1] = Long.parseLong(st.nextToken());
+            arr[i][0] = Integer.parseInt(st.nextToken());
+            arr[i][1] = Integer.parseInt(st.nextToken());
         }
         Arrays.sort(arr, (a, b) -> {
-            if (a[1] == b[1]) return Long.compare(a[0], b[0]);
-            return Long.compare(a[1], b[1]);
+            if (a[1] == b[1]) return Integer.compare(a[0], b[0]);
+            return Integer.compare(a[1], b[1]);
         });
 
         int cnt = 0;
-        long last = 0;
+        int last = 0;
         for (int i = 0; i < N; i++) {
             if (arr[i][0] >= last) {
                 cnt++;
